+++ internal/dremel/read_test.go
@@ -26,7 +26,7 @@ func TestRead(t *testing.T) {
 		{
 			name: "optional and not nested",
 			f:    parse.Field{Type: "Person", TypeName: "*int32", FieldNames: []string{"ID"}, RepetitionTypes: []parse.RepetitionType{parse.Optional}},
-			result: `func readID(x Person) (*int32, int64) {
+			result: `func readID(x Person) (*int32, uint8) {
 	switch {
 	case x.ID == nil:
 		return nil, 0
@@ -45,7 +45,7 @@ func TestRead(t *testing.T) {
 		{
 			name: "optional and nested",
 			f:    parse.Field{Type: "Person", TypeName: "*int32", FieldNames: []string{"Hobby", "Difficulty"}, FieldTypes: []string{"Hobby", "int32"}, RepetitionTypes: []parse.RepetitionType{parse.Optional, parse.Optional}},
-			result: `func readHobbyDifficulty(x Person) (*int32, int64) {
+			result: `func readHobbyDifficulty(x Person) (*int32, uint8) {
 	switch {
 	case x.Hobby == nil:
 		return nil, 0
@@ -59,7 +59,7 @@ func TestRead(t *testing.T) {
 		{
 			name: "mix of optional and required and nested",
 			f:    parse.Field{Type: "Person", TypeName: "string", FieldNames: []string{"Hobby", "Name"}, FieldTypes: []string{"Hobby", "string"}, RepetitionTypes: []parse.RepetitionType{parse.Optional, parse.Required}},
-			result: `func readHobbyName(x Person) (*string, int64) {
+			result: `func readHobbyName(x Person) (*string, uint8) {
 	switch {
 	case x.Hobby == nil:
 		return nil, 0
@@ -71,7 +71,7 @@ func TestRead(t *testing.T) {
 		{
 			name: "mix of optional and required and nested v2",
 			f:    parse.Field{Type: "Person", TypeName: "*string", FieldNames: []string{"Hobby", "Name"}, FieldTypes: []string{"Hobby", "string"}, RepetitionTypes: []parse.RepetitionType{parse.Required, parse.Optional}},
-			result: `func readHobbyName(x Person) (*string, int64) {
+			result: `func readHobbyName(x Person) (*string, uint8) {
 	switch {
 	case x.Hobby.Name == nil:
 		return nil, 0
@@ -83,7 +83,7 @@ func TestRead(t *testing.T) {
 		{
 			name: "mix of optional and require and nested 3 deep",
 			f:    parse.Field{Type: "Person", TypeName: "*string", FieldNames: []string{"Friend", "Hobby", "Name"}, FieldTypes: []string{"Entity", "Item", "string"}, RepetitionTypes: []parse.RepetitionType{parse.Optional, parse.Required, parse.Optional}},
-			result: `func readFriendHobbyName(x Person) (*string, int64) {
+			result: `func readFriendHobbyName(x Person) (*string, uint8) {
 	switch {
 	case x.Friend == nil:
 		return nil, 0
@@ -97,7 +97,7 @@ func TestRead(t *testing.T) {
 		{
 			name: "mix of optional and require and nested 3 deep v2",
 			f:    parse.Field{Type: "Person", TypeName: "*string", FieldNames: []string{"Friend", "Hobby", "Name"}, FieldTypes: []string{"Entity", "Item", "string"}, RepetitionTypes: []parse.RepetitionType{parse.Required, parse.Optional, parse.Optional}},
-			result: `func readFriendHobbyName(x Person) (*string, int64) {
+			result: `func readFriendHobbyName(x Person) (*string, uint8) {
 	switch {
 	case x.Friend.Hobby == nil:
 		return nil, 0
@@ -111,7 +111,7 @@ func TestRead(t *testing.T) {
 		{
 			name: "mix of optional and require and nested 3 deep v3",
 			f:    parse.Field{Type: "Person", TypeName: "string", FieldNames: []string{"Friend", "Hobby", "Name"}, FieldTypes: []string{"Entity", "Item", "string"}, RepetitionTypes: []parse.RepetitionType{parse.Optional, parse.Optional, parse.Required}},
-			result: `func readFriendHobbyName(x Person) (*string, int64) {
+			result: `func readFriendHobbyName(x Person) (*string, uint8) {
 	switch {
 	case x.Friend == nil:
 		return nil, 0
@@ -125,7 +125,7 @@ func TestRead(t *testing.T) {
 		{
 			name: "nested 3 deep all optional",
 			f:    parse.Field{Type: "Person", TypeName: "*string", FieldNames: []string{"Friend", "Hobby", "Name"}, FieldTypes: []string{"Entity", "Item", "string"}, RepetitionTypes: []parse.RepetitionType{parse.Optional, parse.Optional, parse.Optional}},
-			result: `func readFriendHobbyName(x Person) (*string, int64) {
+			result: `func readFriendHobbyName(x Person) (*string, uint8) {
 	switch {
 	case x.Friend == nil:
 		return nil, 0
@@ -141,7 +141,7 @@ func TestRead(t *testing.T) {
 		{
 			name: "four deep",
 			f:    parse.Field{Type: "Person", TypeName: "*string", FieldNames: []string{"Friend", "Hobby", "Name", "First"}, FieldTypes: []string{"Entity", "Item", "Name", "string"}, RepetitionTypes: []parse.RepetitionType{parse.Optional, parse.Optional, parse.Optional, parse.Optional}},
-			result: `func readFriendHobbyNameFirst(x Person) (*string, int64) {
+			result: `func readFriendHobbyNameFirst(x Person) (*string, uint8) {
 	switch {
 	case x.Friend == nil:
 		return nil, 0
@@ -159,7 +159,7 @@ func TestRead(t *testing.T) {
 		{
 			name: "four deep mixed",
 			f:    parse.Field{Type: "Person", TypeName: "*string", FieldNames: []string{"Friend", "Hobby", "Name", "First"}, FieldTypes: []string{"Entity", "Item", "Name", "string"}, RepetitionTypes: []parse.RepetitionType{parse.Required, parse.Optional, parse.Optional, parse.Optional}},
-			result: `func readFriendHobbyNameFirst(x Person) (*string, int64) {
+			result: `func readFriendHobbyNameFirst(x Person) (*string, uint8) {
 	switch {
 	case x.Friend.Hobby == nil:
 		return nil, 0
@@ -175,7 +175,7 @@ func TestRead(t *testing.T) {
 		{
 			name: "four deep mixed v2",
 			f:    parse.Field{Type: "Person", TypeName: "string", FieldNames: []string{"Friend", "Hobby", "Name", "First"}, FieldTypes: []string{"Entity", "Item", "Name", "string"}, RepetitionTypes: []parse.RepetitionType{parse.Optional, parse.Optional, parse.Optional, parse.Required}},
-			result: `func readFriendHobbyNameFirst(x Person) (*string, int64) {
+			result: `func readFriendHobbyNameFirst(x Person) (*string, uint8) {
 	switch {
 	case x.Friend == nil:
 		return nil, 0
